# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Doc API",
        "version": "1.0.0"
    },
    "paths": {
        "/": {},
        "/api/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Регистрация пользователя",
                "description": "Регистрация пользователя и получение токена",
                "operationId": "c47411f6de96b5b30018c647e449e470",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Tom"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "tom@mail.com"
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": "12345678"
                                            },
                                            "password_confirmation": {
                                                "type": "string",
                                                "example": "12345678"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Пользователь зарегистрирован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "success": {
                                                    "type": "boolean",
                                                    "example": "true"
                                                },
                                                "data": {
                                                    "properties": {
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Tom"
                                                        },
                                                        "token": {
                                                            "type": "string",
                                                            "example": "2|cNYrI1xzLYPw7CW56XaTASlTHSCAE1GWwD1x7qoq8b44e56w"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Пользователь зарегистрирован"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Авторизация пользователя",
                "description": "Авторизация пользователя и получение токена",
                "operationId": "82d74137970376a374510566964c9813",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "email": {
                                                "type": "string",
                                                "example": "tom@mail.com"
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": "12345678"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Пользователь вошел в систему",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "success": {
                                                    "type": "boolean",
                                                    "example": "true"
                                                },
                                                "data": {
                                                    "properties": {
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Tom"
                                                        },
                                                        "token": {
                                                            "type": "string",
                                                            "example": "2|cNYrI1xzLYPw7CW56XaTASlTHSCAE1GWwD1x7qoq8b44e56w"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Пользователь вошел в систему!"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/companies": {
            "get": {
                "tags": [
                    "Company"
                ],
                "summary": "Список компаний",
                "description": "Список компаний авторизованного пользователя",
                "operationId": "bac02e71f0b83ba2ee30d9bbe17fd42a",
                "responses": {
                    "200": {
                        "description": "Получение списка компаний авторизованного пользователя",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "success": {
                                                    "type": "boolean",
                                                    "example": "true"
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": "10"
                                                            },
                                                            "inn": {
                                                                "type": "string",
                                                                "example": "7707083893"
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "ПАО СБЕРБАНК"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Список компаний пользователя с ID = 1"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Company"
                ],
                "summary": "Создание компании",
                "description": "Создание компании по ИНН авторизованным пользователем",
                "operationId": "bcd13c72f54074500df5cf4689c42cda",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "inn": {
                                                "type": "string",
                                                "example": "7707083893"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Company created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "success": {
                                                    "type": "boolean",
                                                    "example": "true"
                                                },
                                                "data": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": "10"
                                                        },
                                                        "inn": {
                                                            "type": "string",
                                                            "example": "7707083893"
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "ПАО СБЕРБАНК"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Компания была создана для пользователя с ID = 1"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "User",
            "description": "User"
        },
        {
            "name": "Company",
            "description": "Company"
        }
    ]
}